% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzylm.R
\name{fuzzylm}
\alias{fuzzylm}
\title{Fuzzy Linear Regression}
\usage{
fuzzylm(formula, data, method = "lee", fuzzy.left.x = NULL,
  fuzzy.right.x = NULL, fuzzy.left.y = NULL, fuzzy.right.y = NULL, ...)
}
\arguments{
\item{formula}{a model formula.}

\item{data}{a data.frame, containing the variables in formula.}

\item{method}{method for fitting of the fuzzy linear model. 
Accepts any non-redundant abbreviation.}

\item{fuzzy.left.x}{character string vector specifying column name(s) with the left 
spread of the fuzzy independent variable(s).}

\item{fuzzy.right.x}{character string vector specifying column name(s) with the right 
spread of the fuzzy independent variable(s).}

\item{fuzzy.left.y}{character string vector specifying column name(s) with the left 
spread of the fuzzy dependent variable(s).}

\item{fuzzy.right.y}{character string vector specifying column name(s) with the right 
spread of the fuzzy dependent variable(s).}

\item{...}{additional parameters used by specific methods.}
}
\value{
Returns a \code{fuzzylm} object that includes the model coefficients, limits
  for data predictions from the model and the input data.
}
\description{
A wrapper function that calculates fuzzy regression coeficients using a chosen method.
}
\details{
The implemented methods include \code{\link{lee}} for fitting the fuzzy linear
  regression from the crisp input data (Lee and Tanaka 1999), and \code{\link{diamond}} 
  (Diamond 1988), \code{\link{hung}} (Hung and Yang 2006), \code{\link{nasrabadi}}
  (Nasrabadi et al. 2005) and \code{\link{tanaka}} (Tanaka et al. 1989) methods for
  triangular fuzzy numbers.
}
\examples{
data(fuzzydat)
fuzzylm(y ~ x, data = fuzzydat$lee, method = "lee")
\dontrun{
# returns error due to incorrect number of spreads
fuzzylm(y ~ x, data = fuzzydat$dia, method = "diamond", fuzzy.left.y = "yl")}
# use the same column name for left and right spread, when the method requests 
# non-symmetric fuzzy numbers
fuzzylm(y ~ x, data = fuzzydat$dia, method = "diamond", fuzzy.left.y = "yl", fuzzy.right.y = "yl")
}
\references{
Diamond, P. (1988) Fuzzy least squares. \emph{Information Sciences}
  46(3): 141-157.

  Hung, W.-L. and Yang, M.-S. (2006) An omission approach for detecting 
  outliers in fuzzy regression models. \emph{Fuzzy Sets and Systems} 157: 3109-3122.

  Lee, H. and Tanaka, H. (1999) Fuzzy approximations with non-symmetric fuzzy
  parameters in fuzzy regression analysis. \emph{Journal of the Operations Research
  Society Japan} 42: 98-112.

  Nasrabadi, M. M., Nasrabadi, E. and Nasrabady, A. R. (2005) Fuzzy linear 
  regression analysis: a multi-objective programming approach. \emph{Applied Mathematics
  and Computation} 163: 245-251.

  Tanaka H., Hayashi I. and Watada J. (1989) Possibilistic linear 
  regression analysis for fuzzy data. \emph{European Journal of Operational 
  Research} 40: 389-396.
}
\seealso{
\code{\link{plot.fuzzylm}}, \code{\link{predict.fuzzylm}}, 
  \code{\link{summary.fuzzylm}}
}
